Multi cursor: wire up implementation
