+++ service/history/configs/config.go
@@ -86,7 +86,8 @@ type Config struct {
 	TimerTaskBatchSize                               dynamicconfig.IntPropertyFn
 	TimerTaskWorkerCount                             dynamicconfig.IntPropertyFn
 	TimerTaskMaxRetryCount                           dynamicconfig.IntPropertyFn
-	TimerProcessorEnableSingleCursor                 dynamicconfig.BoolPropertyFn
+	TimerProcessorEnableSingleProcessor              dynamicconfig.BoolPropertyFn
+	TimerProcessorEnableMultiCursor                  dynamicconfig.BoolPropertyFn
 	TimerProcessorEnablePriorityTaskScheduler        dynamicconfig.BoolPropertyFn
 	TimerProcessorSchedulerWorkerCount               dynamicconfig.IntPropertyFn
 	TimerProcessorSchedulerQueueSize                 dynamicconfig.IntPropertyFn
@@ -111,7 +112,8 @@ type Config struct {
 	TransferTaskBatchSize                               dynamicconfig.IntPropertyFn
 	TransferTaskWorkerCount                             dynamicconfig.IntPropertyFn
 	TransferTaskMaxRetryCount                           dynamicconfig.IntPropertyFn
-	TransferProcessorEnableSingleCursor                 dynamicconfig.BoolPropertyFn
+	TransferProcessorEnableSingleProcessor              dynamicconfig.BoolPropertyFn
+	TransferProcessorEnableMultiCursor                  dynamicconfig.BoolPropertyFn
 	TransferProcessorEnablePriorityTaskScheduler        dynamicconfig.BoolPropertyFn
 	TransferProcessorSchedulerWorkerCount               dynamicconfig.IntPropertyFn
 	TransferProcessorSchedulerQueueSize                 dynamicconfig.IntPropertyFn
@@ -235,12 +237,12 @@ type Config struct {
 	VisibilityTaskBatchSize                               dynamicconfig.IntPropertyFn
 	VisibilityTaskWorkerCount                             dynamicconfig.IntPropertyFn
 	VisibilityTaskMaxRetryCount                           dynamicconfig.IntPropertyFn
+	VisibilityProcessorEnableMultiCursor                  dynamicconfig.BoolPropertyFn
 	VisibilityProcessorEnablePriorityTaskScheduler        dynamicconfig.BoolPropertyFn
 	VisibilityProcessorSchedulerWorkerCount               dynamicconfig.IntPropertyFn
 	VisibilityProcessorSchedulerQueueSize                 dynamicconfig.IntPropertyFn
 	VisibilityProcessorSchedulerRoundRobinWeights         dynamicconfig.MapPropertyFn
 	VisibilityProcessorCompleteTaskFailureRetryCount      dynamicconfig.IntPropertyFn
-	VisibilityProcessorFailoverMaxPollRPS                 dynamicconfig.IntPropertyFn
 	VisibilityProcessorMaxPollRPS                         dynamicconfig.IntPropertyFn
 	VisibilityProcessorMaxPollHostRPS                     dynamicconfig.IntPropertyFn
 	VisibilityProcessorMaxPollInterval                    dynamicconfig.DurationPropertyFn
@@ -311,7 +313,8 @@ func NewConfig(dc *dynamicconfig.Collection, numberOfShards int32, isAdvancedVis
 		TimerTaskBatchSize:                               dc.GetIntProperty(dynamicconfig.TimerTaskBatchSize, 100),
 		TimerTaskWorkerCount:                             dc.GetIntProperty(dynamicconfig.TimerTaskWorkerCount, 10),
 		TimerTaskMaxRetryCount:                           dc.GetIntProperty(dynamicconfig.TimerTaskMaxRetryCount, 20),
-		TimerProcessorEnableSingleCursor:                 dc.GetBoolProperty(dynamicconfig.TimerProcessorEnableSingleCursor, false),
+		TimerProcessorEnableSingleProcessor:              dc.GetBoolProperty(dynamicconfig.TimerProcessorEnableSingleProcessor, false),
+		TimerProcessorEnableMultiCursor:                  dc.GetBoolProperty(dynamicconfig.TimerProcessorEnableMultiCursor, false),
 		TimerProcessorEnablePriorityTaskScheduler:        dc.GetBoolProperty(dynamicconfig.TimerProcessorEnablePriorityTaskScheduler, false),
 		TimerProcessorSchedulerWorkerCount:               dc.GetIntProperty(dynamicconfig.TimerProcessorSchedulerWorkerCount, 200),
 		TimerProcessorSchedulerQueueSize:                 dc.GetIntProperty(dynamicconfig.TimerProcessorSchedulerQueueSize, 10),
@@ -335,7 +338,8 @@ func NewConfig(dc *dynamicconfig.Collection, numberOfShards int32, isAdvancedVis
 		TransferTaskBatchSize:                               dc.GetIntProperty(dynamicconfig.TransferTaskBatchSize, 100),
 		TransferTaskWorkerCount:                             dc.GetIntProperty(dynamicconfig.TransferTaskWorkerCount, 10),
 		TransferTaskMaxRetryCount:                           dc.GetIntProperty(dynamicconfig.TransferTaskMaxRetryCount, 20),
-		TransferProcessorEnableSingleCursor:                 dc.GetBoolProperty(dynamicconfig.TransferProcessorEnableSingleCursor, false),
+		TransferProcessorEnableSingleProcessor:              dc.GetBoolProperty(dynamicconfig.TransferProcessorEnableSingleProcessor, false),
+		TransferProcessorEnableMultiCursor:                  dc.GetBoolProperty(dynamicconfig.TransferProcessorEnableMultiCursor, false),
 		TransferProcessorEnablePriorityTaskScheduler:        dc.GetBoolProperty(dynamicconfig.TransferProcessorEnablePriorityTaskScheduler, false),
 		TransferProcessorSchedulerWorkerCount:               dc.GetIntProperty(dynamicconfig.TransferProcessorSchedulerWorkerCount, 200),
 		TransferProcessorSchedulerQueueSize:                 dc.GetIntProperty(dynamicconfig.TransferProcessorSchedulerQueueSize, 10),
@@ -429,11 +433,11 @@ func NewConfig(dc *dynamicconfig.Collection, numberOfShards int32, isAdvancedVis
 		// ===== Visibility related =====
 		VisibilityTaskHighPriorityRPS:                         dc.GetIntPropertyFilteredByNamespace(dynamicconfig.VisibilityTaskHighPriorityRPS, 500),
 		VisibilityTaskBatchSize:                               dc.GetIntProperty(dynamicconfig.VisibilityTaskBatchSize, 100),
-		VisibilityProcessorFailoverMaxPollRPS:                 dc.GetIntProperty(dynamicconfig.VisibilityProcessorFailoverMaxPollRPS, 1),
 		VisibilityProcessorMaxPollRPS:                         dc.GetIntProperty(dynamicconfig.VisibilityProcessorMaxPollRPS, 20),
 		VisibilityProcessorMaxPollHostRPS:                     dc.GetIntProperty(dynamicconfig.VisibilityProcessorMaxPollHostRPS, 0),
 		VisibilityTaskWorkerCount:                             dc.GetIntProperty(dynamicconfig.VisibilityTaskWorkerCount, 10),
 		VisibilityTaskMaxRetryCount:                           dc.GetIntProperty(dynamicconfig.VisibilityTaskMaxRetryCount, 20),
+		VisibilityProcessorEnableMultiCursor:                  dc.GetBoolProperty(dynamicconfig.VisibilityProcessorEnableMultiCursor, false),
 		VisibilityProcessorEnablePriorityTaskScheduler:        dc.GetBoolProperty(dynamicconfig.VisibilityProcessorEnablePriorityTaskScheduler, false),
 		VisibilityProcessorSchedulerWorkerCount:               dc.GetIntProperty(dynamicconfig.VisibilityProcessorSchedulerWorkerCount, 200),
 		VisibilityProcessorSchedulerQueueSize:                 dc.GetIntProperty(dynamicconfig.VisibilityProcessorSchedulerQueueSize, 10),
