+++ service/history/tasks/predicates.go
@@ -103,3 +103,58 @@ func (t *TypePredicate) Equals(predicate Predicate) bool {
 
 	return maps.Equal(t.Types, typePrediate.Types)
 }
+
+func AndPredicates(a Predicate, b Predicate) Predicate {
+	switch a := a.(type) {
+	case *NamespacePredicate:
+		if b, ok := b.(*NamespacePredicate); ok {
+			return &NamespacePredicate{
+				NamespaceIDs: intersect(a.NamespaceIDs, b.NamespaceIDs),
+			}
+		}
+	case *TypePredicate:
+		if b, ok := b.(*TypePredicate); ok {
+			return &TypePredicate{
+				Types: intersect(a.Types, b.Types),
+			}
+		}
+	}
+
+	return predicates.And(a, b)
+}
+
+func OrPredicates(a Predicate, b Predicate) Predicate {
+	switch a := a.(type) {
+	case *NamespacePredicate:
+		if b, ok := b.(*NamespacePredicate); ok {
+			return &NamespacePredicate{
+				NamespaceIDs: union(a.NamespaceIDs, b.NamespaceIDs),
+			}
+		}
+	case *TypePredicate:
+		if b, ok := b.(*TypePredicate); ok {
+			return &TypePredicate{
+				Types: union(a.Types, b.Types),
+			}
+		}
+	}
+
+	return predicates.Or(a, b)
+}
+
+func intersect[K comparable](this, that map[K]struct{}) map[K]struct{} {
+	intersection := make(map[K]struct{})
+	for key := range this {
+		if _, ok := that[key]; ok {
+			intersection[key] = struct{}{}
+		}
+	}
+	return intersection
+}
+
+func union[K comparable](this, that map[K]struct{}) map[K]struct{} {
+	union := make(map[K]struct{}, len(this)+len(that))
+	maps.Copy(union, this)
+	maps.Copy(union, that)
+	return union
+}
